+++ tests/test_cli.py
@@ -1,20 +1,37 @@
+import hashlib
+from datetime import datetime, timedelta
+
 import nest_asyncio
 import pytest
 from minicli import run
 
-from tests.conftest import RESOURCE_ID
+from tests.conftest import RESOURCE_ID, RESOURCE_URL
 
 pytestmark = pytest.mark.asyncio
 nest_asyncio.apply()
 
 
-async def test_purge_checks(setup_catalog, db, fake_check):
-    await fake_check()
-    await fake_check()
+async def test_analysis_csv(setup_catalog, rmock, catalog_content, db, fake_check, produce_mock):
     check = await fake_check()
-    run("purge_checks", limit=2)
-    res = await db.fetch("SELECT id FROM checks WHERE resource_id = $1", check["resource_id"])
+    url = check["url"]
+    rmock.get(url, status=200, body=catalog_content)
+
+    # Analyse using check_id
+    run("analyse-csv", check_id=str(check["id"]))
+    # Analyse using URL
+    run("analyse-csv", url=RESOURCE_URL)
+
+
+async def test_purge_checks(setup_catalog, db, fake_check):
+    await fake_check(created_at=datetime.now() - timedelta(days=50))
+    await fake_check(created_at=datetime.now() - timedelta(days=30))
+    await fake_check(created_at=datetime.now() - timedelta(days=10))
+    run("purge_checks", retention_days=40)
+    res = await db.fetch("SELECT * FROM checks")
     assert len(res) == 2
+    run("purge_checks", retention_days=20)
+    res = await db.fetch("SELECT * FROM checks")
+    assert len(res) == 1
 
 
 async def test_purge_csv_tables(setup_catalog, db, fake_check):