+++ udata_hydra/schemas/__init__.py
@@ -1,3 +1,4 @@
 # ruff: noqa: F401
 from .check import CheckGroupBy, CheckSchema
 from .resource import ResourceDocumentSchema, ResourceSchema
+from .resource_exception import ResourceExceptionSchema